+++ src/drivers/index.ts
@@ -3,6 +3,8 @@
  * 打印机驱动模块 - 支持多种打印机协议
  */
 
+export * from './contracts';
+
 export { EscPos, type EscPosOptions } from './EscPosDriver';
 
 export { GPrinterDriver, type GPrinterOptions } from './GPrinterDriver';